requirements.txt
aiohttp==3.8.5

aiosignal==1.3.1

anaconda==0.0.1.1

anyio==3.7.1

APScheduler==3.10.3

argon2-cffi==21.3.0

argon2-cffi-bindings==21.2.0

arrow==1.2.3

asgiref==3.6.0

asttokens==2.2.1

async-lru==2.0.3

async-timeout==4.0.3

attrs==23.1.0

autopep8==2.0.2

Babel==2.12.1

backcall==0.2.0

bbox-visualizer==0.1.0

beautifulsoup4==4.12.2

bleach==6.0.0

certifi==2023.5.7

cffi==1.15.1

chardet==3.0.4

charset-normalizer==3.1.0

colorama==0.4.6

comm==0.1.3

comtypes==1.2.0

contourpy==1.1.0

cryptography==41.0.3

cycler==0.11.0

debugpy==1.6.7

decorator==5.1.1

defusedxml==0.7.1

Django==4.2.1

django-apscheduler==0.6.2

django-cors-headers==4.1.0

django-crontab==0.6.0

django-filter==23.2

django-phonenumber-field==7.1.0

django-simple-history==3.3.0

django-twilio==0.14.3.2

djangorestframework==3.14.0

executing==1.2.0

fastjsonschema==2.18.0

filelock==3.12.2

fonttools==4.41.1

fqdn==1.5.1

frozenlist==1.4.0

geographiclib==2.0

geopy==2.3.0

googlemaps==4.10.0

googletrans==4.0.0rc1

gunicorn==21.2.0

h11==0.9.0

h2==3.2.0

hpack==3.0.0

hstspreload==2023.1.1

httpcore==0.9.1

httpx==0.13.3

hyperframe==5.2.0

idna==2.10

ipykernel==6.24.0

ipython==8.14.0

ipywidgets==8.1.0

isoduration==20.11.0

jedi==0.18.2

Jinja2==3.1.2

joblib==1.3.1

json5==0.9.14

jsonpointer==2.4

jsonschema==4.18.4

jsonschema-specifications==2023.7.1

jupyter-bbox-widget==0.5.0

jupyter-events==0.6.3

jupyter-lsp==2.2.0

jupyter_client==8.3.0

jupyter_core==5.3.1

jupyter_server==2.7.0

jupyter_server_terminals==0.4.4

jupyterlab==4.0.3

jupyterlab-pygments==0.2.2

jupyterlab-widgets==3.0.8

jupyterlab_server==2.23.0

kiwisolver==1.4.4

lxml==4.9.2

MarkupSafe==2.1.3

matplotlib==3.7.2

matplotlib-inline==0.1.6

mistune==3.0.1

mpmath==1.3.0

multidict==6.0.4

mysqlclient==2.2.0

nbclient==0.8.0

nbconvert==7.7.2

nbformat==5.9.1

nest-asyncio==1.5.6

networkx==3.1

notebook_shim==0.2.3

numpy==1.25.0

openai==0.28.0

opencv-python==4.8.0.74

overrides==7.3.1

packaging==23.1

pandas==2.0.3

pandocfilters==1.5.0

parso==0.8.3

phonenumbers==8.13.15

pickleshare==0.7.5

Pillow==10.0.0

platformdirs==3.8.0

prometheus-client==0.17.1

prompt-toolkit==3.0.39

psutil==5.9.5

psycopg2==2.9.7

pure-eval==0.2.2

py-cpuinfo==9.0.0

PyAudio==0.2.13

pycodestyle==2.10.0

pycparser==2.21

pycryptodome==3.18.0

pydub==0.25.1

Pygments==2.15.1

PyJWT==2.7.0

pylabel==0.1.53

pyparsing==3.0.9

pypiwin32==223

PyQt5==5.15.9

PyQt5-Qt5==5.15.2

PyQt5-sip==12.12.1

python-dateutil==2.8.2

python-json-logger==2.0.7

pyttsx3==2.90

pytz==2023.3

pywin32==306

pywinpty==2.0.11

PyYAML==6.0.1

pyzmq==25.1.0

referencing==0.30.0

requests==2.31.0

rfc3339-validator==0.1.4

rfc3986==1.5.0

rfc3986-validator==0.1.1

rpds-py==0.9.2

scikit-learn==1.3.0

scipy==1.11.1

seaborn==0.12.2

Send2Trash==1.8.2

six==1.16.0

sklearn==0.0.post7

sniffio==1.3.0

soupsieve==2.4.1

SpeechRecognition==3.10.0

sqlparse==0.4.4

stack-data==0.6.2

sympy==1.12

terminado==0.17.1

threadpoolctl==3.2.0

tinycss2==1.2.1

torch==2.0.1

torchvision==0.15.2

tornado==6.3.2

tqdm==4.66.1

traitlets==5.9.0

twilio==7.17.0

typing_extensions==4.7.1

tzdata==2023.3

tzlocal==5.0.1

ultralytics==8.0.163

uri-template==1.3.0

urllib3==2.0.3

utils==1.0.1

wcwidth==0.2.6

webcolors==1.13

webencodings==0.5.1

websocket-client==1.6.1

wget==3.2

whitenoise==6.5.0

widgetsnbextension==4.0.8

word2number==1.1

yarl==1.9.2